renamed iterfaces dir to model
