--- src/app/interfaces/NavBarElement.ts
@@ -1,3 +1,0 @@
-export interface NavBarElement{
-    name: string;
-}
