+++ src/app/views/admin-panel/admin-panel.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { NavBarElement } from 'src/app/interfaces/NavBarElement';
+import { NavBarElement } from 'src/app/model/NavBarElement';
 
 @Component({
   selector: 'app-admin-panel',
