+++ src/app/views/admin-panel/admin-panel.module.ts
@@ -5,9 +5,6 @@ import { ListboxModule } from 'primeng/listbox';
 import { RouterModule } from '@angular/router';
 import { FormsModule } from '@angular/forms';
 
-
-
-
 @NgModule({
   declarations: [
     AdminPanelComponent
